Move dir for db
