+++ README.md
@@ -160,11 +160,14 @@ $ go mod download
 
 ## ディレクトリ構成、設計
 
-- db: DB用の設定ファイルやマイグレーションスクリプトを配置する
+- infra
+  - sqlhandler: DBとのコネクションを確立する
+  - その他: マイグレーションスクリプトを配置する
 - model: データベースに保管されたデータをプログラム上においてどのようなデータ構造で利用するかを記述する
 - repository: データストアからデータを取得するための処理を記述する。クエリ結果をモデルに(GoのStruct)にマッピングし、モデルを返す
-- handler: ハンドラ関数(リクエストを受け取りレスポンスを返す関数)を記述する
-- server: リクエストとハンドラのマッピングをする(ルーティングの役割)
+- handler
+  - router: リクエストとハンドラのマッピングをする(ルーティングの役割)
+  - handler: ハンドラ関数(リクエストを受け取りレスポンスを返す関数)を記述する
 - middleware: カスタムのミドルウェアを設定する
 - util: ドメインモデルとは関連のないヘルパー関数を配置する
